interim1:  Add simple presentation skeleton for first interim presentation
This is a start for our presentation on next Tuesday. Added slightly
adapted fu-beamer-template.tex as layout and basic presentation.tex
containing only 4 frames which indicate the outline of the presentation.
Somehow, page margins are not as desired, need to find out how to change
them.
Lbnl, this introduces our new polygons logo! See logo_inkscape.svg for
source (polygons.eps is the same). Cheers!