PermuteAndReject: First naive implementation of Held's permute & reject generator
This patch adds the PermuteAndReject generator, which naively generates a random
permutation of n points and checks whether the resulting polygon is simple. If not,
it rejects the polygon and starts again. The generator itself, though being really
simple, is untested as I wasn't aware of a way to reasonably verify the output without
having a decent GUI...

Regarding the framework, I decided to go for a simple PGenerator interface myself. In
case you guys want to improve it with the ideas discussed on Tuesday, please go ahead.
The current implementation provides a simple Map<String, Object> as generator parameters.
Parameters are documented in PGenerator.java.
Besides, I removed the history from class Polygon and moved it as a second parameter to
the generator's run method.

Various other points:
* new Edge class describing a line segment with two Points.
* Point has now an equal() method.
* new util package containing MathUtils.java.
* MathUtils contains a simple method to generate n random points on a s*s square.
* Polygon can check its own simplicity using a naive n^2 approach (check each line
  against every other line for intersections).
* Edge can check for intersections/coincidence with other edges. I'm rather unsure
  about this one, as I had to adapt the algorithm a little bit to work for the
  simplicity check in Polygon. Please take a look at Edge.java and comment.
* new tests package for unit tests.