TwoOptMoves: Implemented 2-Opt moves algorithm with bug
* Introduced Polygon interface and GraphPolygon implementation
* 2-opt moves uses the latter
* much more to see...

There's a bug in 2-Opt, which causes new edges to be already present in
edges list and therefore not be added. Edges shrinks in size and only
a few points are left. This also causes getPoints() to fail, as it
expects a roundtripped edges list (vertices with only 2 neighbours,
polygon shaped edges). I'm not sure why this happens, as I presumingly
kept to the paper as close as possible. Maybe I'm creating 2 cycles in the graph
somewhere in the code or maybe the data model still does not work
for 2opt. GraphPolygon::addEdge is the point of failure, see head of GraphPolygon.java
for data model explanation also. Nasty bug.