IC&B: Looking for next random _usable_ point can result in infinite loop. (FIX)
While looking for a random point at the beginning of recursivelyAddPoint(), I wasn't aware
of the fact that there might be no such point (i.e. for all i in remaining,
ue.isMarked(lp, i) is true). Thus, we need to record which points we already tried and backtrack
in case we can't find a usable one.

This fixes the at times occurring infinite loop for bigger n. Remaining problem is that IC&B produces
non-simple polygons at times, which I'm going to dig into this afternoon.