PaintPanel: Reset view on middle click.
I noticed that one can fade out the polygon if one zooms in a bit and afterwards
zooms out a lot. There seems to be an issue with the zoom variable going towards
(or below) zero. Actually, this should have been prevented in mouseWheelMoved.
For now, this patch adds a 'reset view' option, which can be triggered on middle
mouse button click. It resets zoom and offset to default values.