OrderedListPolygon: Fix triangulate().
* Fixed some bugs in triangulate() and isASnip().
* Pulled isASnip back into triangulate().
* Fixed the test case.
* Moved OLP.containsPoint into Polygon and removed containsPoint from
  Triangle.
* Removed the error detection in favor of assert(). We assume ccw-oriented,
  simple polygons.

What is worth mentioning here: In the C++ implementation which Steve
used here there was a line in isSnip() which read like:

if ( EPSILON > (((Bx-Ax)*(Cy-Ay)) - ((By-Ay)*(Cx-Ax))) ) return false;

Steve omitted this line in the Java version completely, which made the
triangulation fail for polygons with concave vertices. I (too?) don't
really understand what they're doing in this line, so I simply added a
orientation test to isASnip, which makes sure the Triangle in consideration
is ccw-oriented (i.e. the snip actually is _inside_ the polygon). I hope
this is equivalent to the math shown above.