Polygon: Fix containsPoint().
There was a strange type in containsPoint, Steve. I found the stackoverflow
article where you presumingly took the code from and it did slightly
different comparison there. Not sure, if it works, as Eclipse fails to
build _anything_ for me right now.

Besides, I gave containsPoint() some more thought and am wondering about
what happens to some special cases, for example when the ray is coincident
with an edge or if it hits a vertex directly. Need to look into it some more.