Seidel: Remember used points instead of used horizontal lines.
A polygon's points may share the same y-coordinate even if they're not
the same points. Remember used points, not horizontal lines. In case we
traverse the search tree to find a point and hit a horizontal line, simply
abort.

This is not really clear in Seidel's paper where he wrote, 'If the upper endpoint
a is not an endpoint of some segment of S already, we first use Z(S) to locate
the trapezoid t_a of T(S) that contains a.' The first part of the sentence DOES
NOT ensure that there really is a trapezoid t_a that contains a. Hopefully this works
now.