ShortestPath: Fix reducePolygon() to work with triangles.
When we hit a concave vertex at either q1 or q2 and the successor of q1
or q2 lies on the same edge as the p_q1 ray polygon intersection, we need
to only return the triangle {p,q1,q2}.

I gave reducePolygon() some more thought and restructured it a little. Should be
correct now.